feat: implement delete and refactoring
