+++ src/main/java/org/hoongoin/interviewbank/account/application/AccountService.java
@@ -28,19 +28,13 @@ import org.hoongoin.interviewbank.account.domain.PasswordResetTokenCommand;
 import org.hoongoin.interviewbank.account.domain.PasswordResetTokenQuery;
 import org.hoongoin.interviewbank.account.application.entity.AccountType;
 import org.hoongoin.interviewbank.account.infrastructure.entity.PasswordResetToken;
-import org.hoongoin.interviewbank.exception.IbBadRequestException;
 import org.hoongoin.interviewbank.exception.IbEntityNotFoundException;
 import org.hoongoin.interviewbank.exception.IbLoginFailedException;
 import org.hoongoin.interviewbank.exception.IbValidationException;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
 
 import lombok.RequiredArgsConstructor;
 
@@ -56,12 +50,6 @@ public class AccountService {
 	private final PasswordResetTokenCommand passwordResetTokenCommand;
 	private final PasswordResetTokenQuery passwordResetTokenQuery;
 	private final MailService mailService;
-	private final AmazonS3 amazonS3;
-
-	@Value("${aws.s3.bucket}")
-	private String bucket;
-	@Value("${aws.region.static}")
-	private String region;
 
 	public RegisterResponse registerByRegisterRequest(RegisterRequest registerRequest) {
 		Account account = accountMapper.registerRequestToAccount(registerRequest);
@@ -131,71 +119,8 @@ public class AccountService {
 	@Transactional
 	public UploadProfileImageResponse saveProfileImage(MultipartFile multipartFile,
 		long requestedAccountId) {
-		byte[] fileBytes = resizeImageSizeOfFile(multipartFile);
-
-		String filename = UUID.randomUUID().toString();
-
-		saveFileToS3(multipartFile, fileBytes, filename);
-
-		String imageUrl = extractS3Url(filename);
-
-		Account account = accountCommandService.updateImageUrl(requestedAccountId, imageUrl);
+		Account account = accountCommandService.updateImageUrl(requestedAccountId, multipartFile);
 
 		return new UploadProfileImageResponse(account.getImageUrl());
-	}
-
-	private byte[] resizeImageSizeOfFile(MultipartFile multipartFile) {
-		BufferedImage originalImage;
-
-		try {
-			originalImage = ImageIO.read(multipartFile.getInputStream());
-		} catch (IOException e) {
-			throw new IbBadRequestException(e.getMessage());
-		}
-
-		int resizedWidth = resize(originalImage.getWidth());
-		int resizedHeight = resize(originalImage.getHeight());
-
-		BufferedImage resizedImage = new BufferedImage(resizedWidth, resizedHeight, originalImage.getType());
-		Graphics2D graphics = resizedImage.createGraphics();
-		graphics.drawImage(originalImage, 0, 0, resizedWidth, resizedHeight, null);
-		graphics.dispose();
-
-		return imageAndContentTypeToBytes(resizedImage, multipartFile.getContentType());
-	}
-
-	private byte[] imageAndContentTypeToBytes(BufferedImage image, String contentType) {
-		try {
-			ByteArrayOutputStream bos = new ByteArrayOutputStream();
-			ImageIO.write(image, getFormatName(contentType), bos);
-			return bos.toByteArray();
-		} catch (IOException e) {
-			throw new IbBadRequestException(e.getMessage());
-		}
-	}
-
-	private String getFormatName(String contentType) {
-		return switch (contentType) {
-			case "image/jpeg" -> "jpeg";
-			case "image/png" -> "png";
-			case "image/gif" -> "gif";
-			default -> throw new IllegalArgumentException("Unsupported image type: " + contentType);
-		};
-	}
-
-	private int resize(int size) {
-		return Math.min(size, 400);
-	}
-
-	private String extractS3Url(String filename) {
-		return "https://" + bucket + ".s3." + region + ".amazonaws.com/" + filename;
-	}
-
-	private void saveFileToS3(MultipartFile multipartFile, byte[] fileBytes,
-		String filename) {
-		ObjectMetadata metadata = new ObjectMetadata();
-		metadata.setContentLength(fileBytes.length);
-		metadata.setContentType(multipartFile.getContentType());
-		amazonS3.putObject(new PutObjectRequest(bucket, filename, new ByteArrayInputStream(fileBytes), metadata));
 	}
 }
